Add basic news sumbmission and voting

+++ Controller/ProductsController.php
@@ -7,8 +7,17 @@ class ProductsController extends AppController {
 		),
 		'Thread' => array(
 			'order' => 'Thread.modified DESC'
+		),
+		'News' => array(
+			'order' => 'News.created DESC',
+			'conditions' => array(
+				'published' => 1,
+				'deleted' => 0
+			),
+			'contain' => array(
+				'User' => array('username', 'slug')
+			)
 		)
-		
 	);
 	
 	public function beforeFilter() {
@@ -24,16 +33,27 @@ class ProductsController extends AppController {
 		}
 		
 		// Get Threads, we need to paginate it
-		// @todo Use 2.0 paginator
 		$threads = $this->paginate('Thread', array('product_id' => $product['Product']['id']));
+		
+		// Paginate News
+		$news = $this->paginate('News', array('product_id' => $product['Product']['id']));
 		
 		// Is this in the users inventory?
 		if($this->userData) {
-			$this->set('inInventory', $this->Product->Inventory->has($product['Product']['id'], $this->userData['id']));
+			$inventory = $this->Product->Inventory->has($product['Product']['id'], AuthComponent::user('id'));
+			if(!empty($inventory)) {
+				$this->request->data = $inventory;
+			}
+			$this->set(compact('inventory'));
 		}
 		
-		$this->set(compact('product', 'category', 'threads'));
+		// Related Products
+		$related = $this->Product->related($product['Product']['id']);
+		
+		$canonical = '/' . $product['Product']['slug'];
+
 		$this->set('title_for_layout', $product['Product']['name'].' &hearts;');
+		$this->set(compact('product', 'category', 'threads', 'canonical', 'related', 'news'));
 	}
 	
 	// Page for users to suggest products for us to cover
@@ -51,37 +71,38 @@ class ProductsController extends AppController {
 			throw new NotFoundException(__('Invalid Product'));
 		}
 		
+		$users = $this->Product->Inventory->haveProduct($product['Product']['id']);
+		
+		$this->set(compact('product', 'users'));
 		
 	}
 	
 	/** Ajax Actions **/
 	public function autocomplete() {
-		if ($this->request->is('ajax')) {
-			$term = $this->request->query['q'];
-			$products = $this->Product->search($term);
-			$categories = $this->Product->search($term);
-			$categories = $this->Product->Category->search($term);
-			
-			$this->set(compact('products', 'categories', 'term'));
-		} else {
-			throw NotFoundException(__('Page not found.'));
+		if (!$this->request->is('ajax')) {
+			throw new MethodNotAllowedException();
 		}
+		
+		$term = $this->request->query['q'];
+		$products = $this->Product->search($term);
+		$categories = $this->Product->Category->search($term);
+		
+		$this->set(compact('products', 'categories', 'term'));
 	}
 	
 	public function top5($categoryId) {
 		// Put this in model
-		if ($this->request->is('ajax')) {
-			$top5Category = $this->Product->Category->find('first',array(
-				'conditions' => array(
-					'Category.id' => $categoryId
-				),
-				array('contain' => false)
-			));
-			$top5Products = $this->Product->topByCategoryId($categoryId);
-			$this->set(compact('top5Category', 'top5Products'));
-		} else {
-			throw NotFoundException(__('Page not found.'));
+		if (!$this->request->is('ajax')) {
+			throw new MethodNotAllowedException();
 		}
+		$top5Category = $this->Product->Category->find('first',array(
+			'conditions' => array(
+				'Category.id' => $categoryId
+			),
+			array('contain' => false)
+		));
+		$top5Products = $this->Product->topByCategoryId($categoryId);
+		$this->set(compact('top5Category', 'top5Products'));
 	}
 	
 	/** Admin Functions **/
@@ -126,11 +147,12 @@ class ProductsController extends AppController {
 		$this->set(compact('categories'));
 	}
 	
-	public function admin_delete($id = null) {
-		if (!$id) {
-			$this->Session->setFlash(__('Invalid id for product', true));
-			$this->redirect(array('action'=>'index'));
+	public function admin_delete($id) {
+		
+		if (!$this->request->is('post')) {
+			throw new MethodNotAllowedException();
 		}
+		
 		if ($this->Product->delete($id)) {
 			$this->Session->setFlash(__('Product deleted', true));
 			$this->redirect(array('action'=>'index'));
